+++ services/web/incoming.js
@@ -11,7 +11,7 @@ function router(dependencies) {
 
   // Find all the API handlers and add them to our Router
   glob
-    .sync(__dirname + "/../incoming/**/*.js")
+    .sync(__dirname + "/../../incoming/**/*.js")
     .map((filename) => require(`${filename}`))
     .forEach((handler) => {
       if (handler.path != null) {
